+++ railties/lib/rails/application/finisher.rb
@@ -97,8 +97,8 @@ module Rails
       end
 
       initializer :activate_queue_consumer do |app|
-        if config.queue == ActiveSupport::Queue
-          app.queue_consumer = config.queue_consumer.start(app.queue, {logger: Rails.logger})
+        if config.queue.class == ActiveSupport::Queue
+          app.queue_consumer = config.queue_consumer.start
           at_exit { app.queue_consumer.shutdown }
         end
       end